+++ src/dom/index.ts
@@ -17,6 +17,7 @@ export {
 export { addEventListener, delegateEvents } from './events'
 export { batch, computed as memo, untracked as untrack } from '@preact/signals-core'
 import { effect as createEffect } from '@preact/signals-core'
+export { clearRange, commitRange } from './helpers'
 
 export function effect<T>(fn: (previous: T) => T, value: T): () => void
 export function effect(fn: () => void): () => void
